#10 refactoring
